FiX: MAD Verified/Unverified.
INFO:
MAD's `calc_endminsec` format is either `null` (unverified) or `MM:SS` (verified) representing the despawn minute and second.

BUG:
`expire_timestamp_verified` is currently assigned using `floatval($pokemon["expire_timestamp_verified"])` which, on a value of 00:17, results in a value of `0` which is later interpreted as unverified (wrong).

FIX:
The actual value doesn't appear to used for anything so assigning a value of `1` (or higher) if verified and `null` (or 0) if unverified solves the issue while remaining compatible with the current logic in `map.js` and without adding a special condition.