Update Map Markers When Receiving Updated IV Information
**BEFORE:**
Skip the received data if already in `mapData.pokemons` without any further checking.

**NOW:**
If already in `mapData.pokemons`, also check if the stored information matches the received information. If there's a difference, remove the marker, delete from dictionary and continue on treating it as a new spawn.

**KNOWN ISSUE:**
Some double notification for the same Pokémon if `Notification->Notify of Pokémon` is enabled and we first receive the encounter without IV (notification sent) and then we receive IV information on a subsequent refresh payload (another notification sent).

Considering the map now remains up-to-date without having to refresh or disable/enable Pokémon and since it's in his `Notify of Pokémon`, I feel it’s an acceptable side-effect and most will be ok with receiving a 2nd notification to see it now has IV information.

**TODO:**
Do we fix the double notifications or leave as-is to inform the user of the updated IV information?