fix pokemon despawn time outside of popup Using MAD since PR#551 has been merged, the despawn time of the popup was offset and partly outside the popup.
This reworks the css to solve the issue and removes the need for the `left:50px`. It also solves the issue on Firefox and Edge where the confirmed checkmark would wrap to the next line.

Appears to be working correctly on Chrome, Firefox and Edge with and without a value in `firstSeen`. Untested using Safari/iOS.