iconsArray logic change for returning users to better handle of all situations
__**ISSUE:-**__
Was not updating String values for existing users or checking if any existing values were still valid and available in the config.

__**CHANGE:**__
`$iconFolderArray`’s values can only be a string or an Array.

If it’s a String, it’s a fixed value that cannot be changed. Always assign the config’s value which will either create or update the key.

If it’s an Array and the key doesn’t already exists, add the key with the config’s first value (default).

If it’s an Array and the key already exists but the value of the existing key is no longer present in the config, update the existing key to the config’s first value (default). This ensures the user is always using a current repo from the config instead of one that could have been removed or changed by the operator.

Hopefully, that now solves any issues associated with that variable.