add with or without AR-Scan Task  toggle for quests (rdm only) RDM: Quick implementation that adds a «With AR-Scan Task» toggle in the Pokéstop menu to allow users to display either the quest when they are holding an AR-Scan task or the quests when they are not holding an AR-Scan task if they are available in your database.
By default, the toggle is disabled and PMSF uses the old behavior of only using with AR quests everywhere. To activate the toggle and let your users decide, set the new variable `$noQuestsARTaskToggle = false;` in your config/access-config.

Search returns results based on the toggle position or with AR when the toggle is disabled.

Pokéstop filters always displays all the available rewards in your database regardless of `$noQuestsARTaskToggle`.

FullStats: Untouched. Still displays only with AR-Scan Task.

MAD:
No changes but things are in place for when the async branch goes mainstream and dual quests are officially supported by mad.