pokemon fetching and processing overhaul - overhaul of the mons fetching, processing and clearing (ALL) - traffic reduction between the user and server when PVP filters are active (RDM) - new And/Or logic available for independent PVP filtering and IV/LVL filtering selectable using a toggle (RDM) - mad will now compute the iv and level in SQL to make both backend uniform and remove extra checks. However, the SQL filtering remains unchanged and still uses the `cp_multiplier` despite its flaws (MAD) - add a blue shadow for PVP Rank 1 mons (RDM) - misc optimisations, cleanup and fixes (ALL)
**New config var:**
`$noIndependantPvpAndStatsToggle`. Default false (Enabled).
`$noPvpShadow`. Default false (Enabled).

PMSF was not initially built to handle pvp. It was built at a time when only iv and level filtering existed and it still worked as intended to this day if pvp filtering was inactive or disabled (or using mad!).

However, with pvp filtering enabled and a normal setting of «Ex: Ranks[3/3/3] + IV & LVL[0/0]» to get Top 3 in all leagues, PMSF would fetch ALL the rows from the database that can be visible on the user’s map,  return all of those to php for formatting, pass all of those to the user’s client which would finally filter them out based on the pvp filters. When zoomed out enough and not hiding any mons, this was a multi megaByte payload sent to the user when it does a full reload (open map, move around, change filters, etc) plus 60+kB for normal refreshes every 5 seconds that often didn’t add anything to the map (all rejected).  Especially for mobile users on quotas, that is extremely bad and inefficient.

Unfortunately, only iv and level filtering can be done directly with a query. There is no method to have a SQL query directly filter by ranks with the way the PVP information is stored in the database. To filter by PVP ranks, all mons that can be visible on the user’s map must be requested, then cycle through leagues to get the best ranks and finally apply the filtering.

If PVP filtering is active and requires all rows, might as well also allow to filter by IV/LVL independently from PVP at the same time since it’s in the data from the SQL query!

Initial filtering will be done with the SQL query whenever possible for best performance. With the query results, filtering will now occur before passing data to the client to allow for greater filtering flexibility and only send relevant information to the client which substantially reduces the server to client traffic when PVP filtering is active.

Because the client no longer has to parse through all that data, it ends up being faster even while also doing PVP and IV/LVL filtering independently! With PVP filtering not active, the performance is similar (it’s the same query as before this commit).